+++ src/routes/stash/+page.svelte
@@ -13,16 +13,15 @@
 	}
 </script>
 
-<div class="flex flex-col w-11/12 md:w-7/12 mx-auto">
+<div class="poem-list flex flex-col w-11/12 md:w-7/12 mx-auto">
 	<div class="overflow-x-auto">
 		<div class="inline-block min-w-full">
 			<div class="overflow-hidden">
 				<table class="min-w-full text-left table-fixed">
 					<tbody>
 						{#if $poems}
-							<tr><th /></tr>
 							{#each $poems as poem (poem.id)}
-								<tr class="border-b bg-white dark:bg-stone-800 dark:text-stone-100 dark:border-b-stone-100">
+								<tr class="border-b dark:bg-stone-800 dark:text-stone-100 dark:border-b-stone-100">
 									<button on:click={openPoem(poem.id)}
 										><td
 											class="w-10/12 whitespace-nowrap py-4 underline decoration-dotted hover:no-underline"