+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## Added
 - New fonts: Arial, Crimson Roman, Comic Sans MS, Times Old Attic Bold
+- Day theme switching
+- Themes: Strawberry Sundae
 
 ## Fixed
 - Poem alignment not being stored in local storage